move combine_serving_endpoint_key to endpoint_mapping.py, lint
