+++ mlrun/serving/endpoint_mapping.py
@@ -27,9 +27,7 @@ import jsonpath_ng.exceptions
 import mlrun.common.schemas as schemas
 import mlrun.errors
 import mlrun.model
-import mlrun.serving.utils as serving_utils
 import mlrun.utils
-
 
 # ---------------------------------------------------------------------------
 # Shared HTTP method validator (used by both EndpointConfig and APIHandlerConfig)
@@ -57,6 +55,11 @@ def _validate_http_method(http_method: HTTPMethod | str) -> HTTPMethod:
         f"http_method must be an HTTPMethod enum or string, got {type(http_method).__name__} "
         f"with value '{http_method}'. Valid values are: {', '.join(m.value for m in HTTPMethod)}"
     )
+
+
+def combine_serving_endpoint_key(method: HTTPMethod, path: str) -> str:
+    """Combine method and path to create a unique endpoint key."""
+    return f"{method.value}:{path}"
 
 
 # ---------------------------------------------------------------------------
@@ -270,7 +273,7 @@ class EndpointConfig(mlrun.model.ModelObj):
 
     def get_endpoint_key(self) -> str:
         """Return the endpoint key in the format 'METHOD:path', e.g. 'POST:/v1/chat/completions'."""
-        return serving_utils.combine_serving_endpoint_key(self.http_method, self.path)
+        return combine_serving_endpoint_key(self.http_method, self.path)
 
     def __repr__(self) -> str:
         return (
@@ -346,7 +349,7 @@ class APIHandlerConfig(mlrun.model.ModelObj):
         """Get endpoint configuration for a specific method and path."""
         method = _validate_http_method(method)
         path = EndpointConfig._normalize_path(path)
-        endpoint_key = serving_utils.combine_serving_endpoint_key(method, path)
+        endpoint_key = combine_serving_endpoint_key(method, path)
         return self._endpoints.get(endpoint_key)
 
     def add_endpoint_handler(
@@ -403,7 +406,7 @@ class APIHandlerConfig(mlrun.model.ModelObj):
         """
         http_method = _validate_http_method(http_method)
         path = EndpointConfig._normalize_path(path)
-        endpoint_key = serving_utils.combine_serving_endpoint_key(http_method, path)
+        endpoint_key = combine_serving_endpoint_key(http_method, path)
         self._endpoints.pop(endpoint_key, None)
 
     def to_dict(self, fields=None, exclude=None, strip=False):
@@ -513,6 +516,57 @@ def compile_dynamic_path_patterns(
     return template_patterns, star_patterns
 
 
+def check_body_and_path_parameters_overlapping(
+    template_patterns: list[tuple[HTTPMethod, Pattern, EndpointConfig]],
+    star_patterns: list[tuple[HTTPMethod, str, EndpointConfig]],
+) -> None:
+    """Check that input_body_mappings destination_path names don't conflict with path
+    template parameter names that would be extracted on the same request.
+
+    Two sources of conflict for each template endpoint:
+    1. Same endpoint — the template endpoint itself has input_body_mappings with a conflicting name.
+    2. Star endpoint — a star endpoint whose prefix covers the template's path has
+       input_body_mappings with a conflicting name (its mappings apply to all requests under
+       its prefix, including requests that also match the template).
+
+    :raises mlrun.errors.MLRunValueError: On config-time conflict detection.
+    """
+    for template_method, compiled_pattern, template_ep in template_patterns:
+        path_param_names = set(compiled_pattern.groupindex.keys())
+
+        # Source 1: same endpoint has input_body_mappings with conflicting destination_path
+        candidates = [(template_ep, "same endpoint")]
+
+        # Source 2: star endpoints whose prefix covers this template's path
+        for star_method, prefix, star_ep in star_patterns:
+            if star_method != template_method:
+                continue
+            if template_ep.path.startswith(prefix):
+                candidates.append(
+                    (star_ep, f"star endpoint '{star_ep.get_endpoint_key()}'")
+                )
+
+        for candidate_ep, source_desc in candidates:
+            if not candidate_ep.input_body_mappings:
+                continue
+            dest_names = {
+                m["destination_path"]
+                for m in candidate_ep.input_body_mappings.mappings
+                if m.get("destination_path")
+            }
+            overlapping = dest_names & path_param_names
+            if overlapping:
+                raise mlrun.errors.MLRunValueError(
+                    f"Configuration conflict: input_body_mappings destination_path(s) "
+                    f"{', '.join(sorted(overlapping))} from {source_desc} "
+                    f"overlap with path template parameter(s) in pattern "
+                    f"'{compiled_pattern.pattern}' "
+                    f"(endpoint '{template_ep.get_endpoint_key()}'). "
+                    f"Rename the destination_path(s) or the path template "
+                    f"placeholder(s) to avoid ambiguity."
+                )
+
+
 def compile_body_map(
     body_mappings: BodyMappings,
     endpoint_key: str,
@@ -570,7 +624,7 @@ def collect_endpoint_matches(
     matches: list[EndpointMatch] = []
 
     # Phase 1: Exact match
-    endpoint_key = serving_utils.combine_serving_endpoint_key(method, path)
+    endpoint_key = combine_serving_endpoint_key(method, path)
     exact_found = endpoint_key in endpoints
     if exact_found:
         matches.append(EndpointMatch(endpoints[endpoint_key]))
